feat(price): rename, also improved annotations
